Adjust IService and IServiceDriver interfaces and related code
See source code for in-detail changes, but the tl;dr is adjusted and uniformed the IService and IServiceDriver interfaces and related code such as class Service  (the implementation of IService) and enum RequestType. Added some new properties such as `isAdverisement`, `isResponseReady` and `awaitResponseReady` and renamed some of the other property and method names. Made `onResponse` signal fire on next event loop, so it will not halt the acception/rejection promise. Added, tweaked and removed some comments. Renamed the exported business logic as `referenceBusinessLogic` instead, as it seems more appropriate.

P.S.: I need to practice writing detailed commit messages...