Make read-only properties read-only at runtime
Previously the properties was only read-only at design time, with
typescript warning us to not set the property, but at design time it
was possible to re-set the properties.

Now the properties are also marked read-only (or non-writable) at
runtime, so the properties are _truly_ read-only, as the runtime will
throw if we try to set the property to a new value (using the normal
way).