Be on the safer side (again)
Don't assume `node-fetch#Response.body` is an instance of
`stream#Readable`. After taking a closer look at the `node-fetch`
source-code I think it is safer to (again) pass the response body
through a `stream#PassThrough` (that _is_ an extension of
`stream#Readable`).

P.s.: Sending a punch to my previous self for assuming things where
we should not.