Remove Request.ip
because I felt it does not belong within the "core" library.

The proxy-example now shows how you _could_ get the ip from the
request if you really need it.

Also changed the signature of Context.state, so it is easier to append
known properties in extending libraries/applications, and tweaked
Context.setHeader to do nothing on types other then expected.